+++ CHANGELOG.md
@@ -87,7 +87,8 @@ And please only add new entries to the top of this list, right below the `# Unre
 - **Breaking:** `with_x11_visual` now takes the visual ID instead of the bare pointer.
 - On X11, add a `with_embedded_parent_window` function to the window builder to allow embedding a window into another window.
 - On iOS, add force data to touch events when using the Apple Pencil.
-- On iOS, add events `Event::Foreground`, `Event::Background` and `Event::MemoryWarning`.
+- On iOS, send events `WindowEvent::Occluded(false)`, `WindowEvent::Occluded(true)` when application enters/leaves foreground.
+- on iOS, add event `Event::MemoryWarning`.
 - On Android, add event `Event::MemoryWarning`.
 
 # 0.29.0-beta.0